refactor(plan): Pull out more planning
