+++ src/plan.rs
@@ -6,6 +6,54 @@ use crate::replace::Template;
 use crate::version::VersionExt as _;
 use crate::*;
 
+pub(crate) fn load(
+    args: &args::ReleaseOpt,
+    ws_meta: &cargo_metadata::Metadata,
+) -> Result<indexmap::IndexMap<cargo_metadata::PackageId, PackageRelease>, error::FatalError> {
+    let root = git::top_level(ws_meta.workspace_root.as_std_path())?;
+
+    let member_ids = cargo::sort_workspace(&ws_meta);
+    member_ids
+        .iter()
+        .filter_map(|p| PackageRelease::load(args, &root, &ws_meta, &ws_meta[p]).transpose())
+        .map(|p| p.map(|p| (p.meta.id.clone(), p)))
+        .collect()
+}
+
+pub(crate) fn plan(
+    mut pkgs: indexmap::IndexMap<cargo_metadata::PackageId, PackageRelease>,
+) -> Result<indexmap::IndexMap<cargo_metadata::PackageId, PackageRelease>, error::FatalError> {
+    let mut shared_max: Option<version::Version> = None;
+    let mut shared_ids = indexmap::IndexSet::new();
+    for (pkg_id, pkg) in pkgs.iter() {
+        if pkg.config.shared_version() {
+            shared_ids.insert(pkg_id.clone());
+            let planned = pkg.version.as_ref().unwrap_or(&pkg.prev_version);
+            if shared_max
+                .as_ref()
+                .map(|max| max.full_version < planned.full_version)
+                .unwrap_or(true)
+            {
+                shared_max = Some(planned.clone());
+            }
+        }
+    }
+    if let Some(shared_max) = shared_max {
+        for shared_id in shared_ids {
+            let shared_pkg = &mut pkgs[&shared_id];
+            if shared_pkg.prev_version.bare_version != shared_max.bare_version {
+                shared_pkg.version = Some(shared_max.clone());
+            }
+        }
+    }
+
+    for pkg in pkgs.values_mut() {
+        pkg.plan()?;
+    }
+
+    Ok(pkgs)
+}
+
 pub struct PackageRelease {
     pub meta: cargo_metadata::Package,
     pub manifest_path: PathBuf,
